Move products to separate route, format imports, correct testid
